requirements.txt
requests
plotly
dash
dash-auth
ruamel.yaml
numpy
numba
pyopenssl
pyyaml
gspread
oauth2client
pprint
ipywidgets